Truncate token uri in logging (#99)
Some URLs are very long (e.g. data urls), so we truncate them to make them easier to read in logs and use less space.

This PR also fixes the foundation parser, which was broken, because the API it relied on doesn't exist anymore. It's now falling back to contract calls.